Fix SDCC error 158: change loop counter from uint8_t to uint16_t
The previous fix of adding 'U' suffix to the constant didn't resolve
the issue. The root cause is that SDCC raises an overflow warning when
comparing a uint8_t variable against its maximum value (255).

Changed the loop counter 'i' from uint8_t to uint16_t in bl_main()
function (line 73). This eliminates the overflow concern since a
uint16_t can safely hold values much larger than 255.

This resolves: bootloader/bootloader.c:102: error 158: overflow in
implicit constant conversion

+++ Firmware/bootloader/bootloader.c
@@ -70,7 +70,7 @@ bool		app_valid;
 void
 bl_main(void)
 {
-	uint8_t		i;
+	uint16_t	i;
 #ifdef FLASH_BANKS
 	uint8_t		bank_state = PSBANK;
 #endif
@@ -99,7 +99,7 @@ bl_main(void)
 	// Do some simple debouncing on the bootloader-entry
 	// strap/button.
 	debounce_count = 0;
-	for (i = 0; i < 255U; i++) {
+	for (i = 0; i < 255; i++) {
 		if (BUTTON_BOOTLOAD == BUTTON_ACTIVE)
 			debounce_count++;
 	}
